+++ src/plugins/intel_gpu/src/plugin/transformations/disable_f16_comp_for_periodic_funcs.hpp
